+++ lib/features/home/home.dart
@@ -1,0 +1,1 @@
+export 'view/view.dart';
