+++ lib/features/home/view/home_page.dart
@@ -1,28 +1,5 @@
-import 'package:flutter/material.dart';
 import 'package:chicken_box_time/flavor/flavor.dart';
-
-void main() {
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: const MyHomePage(title: 'Flutter Demo Home Page'),
-      debugShowCheckedModeBanner: false,
-      builder: (BuildContext context, Widget? child) {
-        return FlavorBanner(child: child);
-      },
-    );
-  }
-}
+import 'package:flutter/material.dart';
 
 class MyHomePage extends StatefulWidget {
   const MyHomePage({Key? key, required this.title}) : super(key: key);
