+++ lib/features/home/view/view.dart
@@ -1,0 +1,1 @@
+export 'home_page.dart';
