+++ lib/main_desktop.dart
@@ -1,4 +1,4 @@
-import 'main.dart' as original_main;
+import 'main_development.dart' as original_main;
 
 // This file is the default main entry-point for go-flutter application.
 void main() {
