+++ lib/main_development.dart
@@ -1,9 +1,9 @@
+import 'package:chicken_box_time/app/app.dart';
 import 'package:chicken_box_time/flavor/flavor.dart';
-import 'package:chicken_box_time/main.dart';
 import 'package:flutter/widgets.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
   Flavor.setEnvironment(Environment.development);
-  runApp(const MyApp());
+  runApp(const App());
 }
